#222 refactor: CarColorType 패키지 이동
