+++ android/MyCarchiving/app/src/main/java/com/softeer/mycarchiving/ui/component/CompleteItems.kt
@@ -34,7 +34,7 @@ import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.skydoves.landscapist.ImageOptions
 import com.skydoves.landscapist.glide.GlideImage
-import com.softeer.data.CarColorType
+import com.softeer.domain.model.CarColorType
 import com.softeer.mycarchiving.R
 import com.softeer.mycarchiving.model.makingcar.ColorOptionUiModel
 import com.softeer.mycarchiving.model.makingcar.SelectOptionUiModel
