+++ android/MyCarchiving/app/src/main/java/com/softeer/mycarchiving/ui/makingcar/complete/CompleteScreen.kt
@@ -11,15 +11,13 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.ViewModelStoreOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.softeer.data.CarColorType
-import com.softeer.mycarchiving.MainActivity
+import com.softeer.domain.model.CarColorType
 import com.softeer.mycarchiving.R
 import com.softeer.mycarchiving.model.makingcar.ColorOptionUiModel
 import com.softeer.mycarchiving.model.makingcar.SelectOptionUiModel
