+++ android/MyCarchiving/data/src/main/java/com/softeer/data/model/TrimType.kt
@@ -1,4 +1,4 @@
-package com.softeer.data
+package com.softeer.data.model
 
 enum class TrimType {
     ENGINE, BODY_TYPE, DRIVING_SYSTEM
