+++ android/MyCarchiving/domain/src/main/java/com/softeer/domain/model/CarColorType.kt
@@ -1,4 +1,4 @@
-package com.softeer.data
+package com.softeer.domain.model
 
 enum class CarColorType(val type: String) {
     EXTERIOR("외장 색상"), INTERIOR("내장 색상")
